+++ frontend/src/components/ui/Logo.tsx
@@ -1,0 +1,28 @@
+export default function Logo() {
+  return (
+    <svg
+      width="109"
+      height="32"
+      viewBox="0 0 109 32"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <g clip-path="url(#clip0_49_113)">
+        <path
+          d="M1.51705 24.5V7.04545H8.40341C9.72727 7.04545 10.8551 7.2983 11.7869 7.80398C12.7188 8.30398 13.429 9 13.9176 9.89204C14.4119 10.7784 14.6591 11.8011 14.6591 12.9602C14.6591 14.1193 14.4091 15.142 13.9091 16.0284C13.4091 16.9148 12.6847 17.6051 11.7358 18.0994C10.7926 18.5937 9.65057 18.8409 8.30966 18.8409H3.92045V15.8835H7.71307C8.4233 15.8835 9.00852 15.7614 9.46875 15.517C9.93466 15.267 10.2813 14.9233 10.5085 14.4858C10.7415 14.0426 10.858 13.5341 10.858 12.9602C10.858 12.3807 10.7415 11.875 10.5085 11.4432C10.2813 11.0057 9.93466 10.6676 9.46875 10.429C9.00284 10.1847 8.41193 10.0625 7.69602 10.0625H5.20739V24.5H1.51705ZM20.3693 24.7472C19.5341 24.7472 18.7898 24.6023 18.1364 24.3125C17.483 24.017 16.9659 23.5824 16.5852 23.0085C16.2102 22.429 16.0227 21.7074 16.0227 20.8438C16.0227 20.1165 16.1563 19.5057 16.4233 19.0114C16.6903 18.517 17.054 18.1193 17.5142 17.8182C17.9744 17.517 18.4972 17.2898 19.0824 17.1364C19.6733 16.983 20.2926 16.875 20.9403 16.8125C21.7017 16.733 22.3153 16.6591 22.7812 16.5909C23.2472 16.517 23.5852 16.4091 23.7955 16.267C24.0057 16.125 24.1108 15.9148 24.1108 15.6364V15.5852C24.1108 15.0455 23.9403 14.6278 23.5994 14.3324C23.2642 14.0369 22.7869 13.8892 22.1676 13.8892C21.5142 13.8892 20.9943 14.0341 20.608 14.3239C20.2216 14.608 19.9659 14.9659 19.8409 15.3977L16.483 15.125C16.6534 14.3295 16.9886 13.642 17.4886 13.0625C17.9886 12.4773 18.6335 12.0284 19.4233 11.7159C20.2188 11.3977 21.1392 11.2386 22.1847 11.2386C22.9119 11.2386 23.608 11.3239 24.2727 11.4943C24.9432 11.6648 25.5369 11.929 26.054 12.2869C26.5767 12.6449 26.9886 13.1051 27.2898 13.6676C27.5909 14.2244 27.7415 14.892 27.7415 15.6705V24.5H24.2983V22.6847H24.196C23.9858 23.0938 23.7045 23.4545 23.3523 23.767C23 24.0739 22.5767 24.3153 22.0824 24.4915C21.5881 24.6619 21.017 24.7472 20.3693 24.7472ZM21.4091 22.2415C21.9432 22.2415 22.4148 22.1364 22.8239 21.9261C23.233 21.7102 23.554 21.4205 23.7869 21.0568C24.0199 20.6932 24.1364 20.2812 24.1364 19.821V18.4318C24.0227 18.5057 23.8665 18.5739 23.6676 18.6364C23.4744 18.6932 23.2557 18.7472 23.0114 18.7983C22.767 18.8437 22.5227 18.8864 22.2784 18.9261C22.0341 18.9602 21.8125 18.9915 21.6136 19.0199C21.1875 19.0824 20.8153 19.1818 20.4972 19.3182C20.179 19.4545 19.9318 19.6392 19.7557 19.8722C19.5795 20.0994 19.4915 20.3835 19.4915 20.7244C19.4915 21.2188 19.6705 21.5966 20.0284 21.858C20.392 22.1136 20.8523 22.2415 21.4091 22.2415ZM30.5582 24.5V11.4091H34.0781V13.6932H34.2145C34.4531 12.8807 34.8537 12.267 35.4162 11.8523C35.9787 11.4318 36.6264 11.2216 37.3594 11.2216C37.5412 11.2216 37.7372 11.233 37.9474 11.2557C38.1577 11.2784 38.3423 11.3097 38.5014 11.3494V14.571C38.331 14.5199 38.0952 14.4744 37.794 14.4347C37.4929 14.3949 37.2173 14.375 36.9673 14.375C36.4332 14.375 35.956 14.4915 35.5355 14.7244C35.1207 14.9517 34.7912 15.2699 34.5469 15.679C34.3082 16.0881 34.1889 16.5597 34.1889 17.0938V24.5H30.5582ZM43.6683 20.733L43.6768 16.3778H44.2053L48.3984 11.4091H52.5661L46.9325 17.9886H46.0717L43.6683 20.733ZM40.3786 24.5V7.04545H44.0092V24.5H40.3786ZM48.5604 24.5L44.7081 18.7983L47.1286 16.233L52.8132 24.5H48.5604Z"
+          fill="white"
+        />
+        <rect x="55" width="54" height="32" rx="3" fill="#F59E0B" />
+        <path
+          d="M64.0795 16.9318V24.5H60.4489V7.04545H63.9773V13.7188H64.1307C64.4261 12.946 64.9034 12.3409 65.5625 11.9034C66.2216 11.4602 67.0483 11.2386 68.0426 11.2386C68.9517 11.2386 69.7443 11.4375 70.4205 11.8352C71.1023 12.2273 71.6307 12.7926 72.0057 13.5312C72.3864 14.2642 72.5739 15.142 72.5682 16.1648V24.5H68.9375V16.8125C68.9432 16.0057 68.7386 15.3778 68.3239 14.929C67.9148 14.4801 67.3409 14.2557 66.6023 14.2557C66.108 14.2557 65.6705 14.3608 65.2898 14.571C64.9148 14.7812 64.6193 15.0881 64.4034 15.4915C64.1932 15.8892 64.0852 16.3693 64.0795 16.9318ZM83.8203 18.9261V11.4091H87.451V24.5H83.9652V22.1222H83.8288C83.5334 22.8892 83.0419 23.5057 82.3544 23.9716C81.6726 24.4375 80.8402 24.6705 79.8572 24.6705C78.9822 24.6705 78.2124 24.4716 77.5476 24.0739C76.8828 23.6761 76.3629 23.1108 75.9879 22.3778C75.6186 21.6449 75.4311 20.767 75.4254 19.7443V11.4091H79.0561V19.0966C79.0618 19.8693 79.2692 20.4801 79.6783 20.929C80.0874 21.3778 80.6357 21.6023 81.3232 21.6023C81.7607 21.6023 82.1697 21.5028 82.5504 21.304C82.9311 21.0994 83.2379 20.7983 83.4709 20.4006C83.7095 20.0028 83.826 19.5114 83.8203 18.9261ZM90.4233 24.5V7.04545H94.054V13.608H94.1648C94.3239 13.2557 94.554 12.8977 94.8551 12.5341C95.1619 12.1648 95.5597 11.858 96.0483 11.6136C96.5426 11.3636 97.1563 11.2386 97.8892 11.2386C98.8438 11.2386 99.7244 11.4886 100.531 11.9886C101.338 12.483 101.983 13.2301 102.466 14.2301C102.949 15.2244 103.19 16.4716 103.19 17.9716C103.19 19.4318 102.955 20.6648 102.483 21.6705C102.017 22.6705 101.381 23.429 100.574 23.946C99.7727 24.4574 98.875 24.7131 97.8807 24.7131C97.1761 24.7131 96.5767 24.5966 96.0824 24.3636C95.5938 24.1307 95.1932 23.8381 94.8807 23.4858C94.5682 23.1278 94.3295 22.767 94.1648 22.4034H94.0028V24.5H90.4233ZM93.9773 17.9545C93.9773 18.733 94.0852 19.4119 94.3011 19.9915C94.517 20.571 94.8295 21.0227 95.2386 21.3466C95.6477 21.6648 96.1449 21.8239 96.7301 21.8239C97.321 21.8239 97.821 21.6619 98.2301 21.3381C98.6392 21.0085 98.9489 20.554 99.1591 19.9744C99.375 19.3892 99.483 18.7159 99.483 17.9545C99.483 17.1989 99.3778 16.5341 99.1676 15.9602C98.9574 15.3864 98.6477 14.9375 98.2386 14.6136C97.8295 14.2898 97.3267 14.1278 96.7301 14.1278C96.1392 14.1278 95.6392 14.2841 95.2301 14.5966C94.8267 14.9091 94.517 15.3523 94.3011 15.9261C94.0852 16.5 93.9773 17.1761 93.9773 17.9545Z"
+          fill="white"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_49_113">
+          <rect width="109" height="32" fill="white" />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+}
